+++ API_BANKING_PAYMENT/Respositories/IRepositories/IBankRepository.cs
@@ -5,8 +5,10 @@ namespace API_BANKING_PAYMENT.Respositories.IRepositories
 {
     public interface IBankRepository : IRepository<Bank>
     {
-        Task<Bank> GetBankWithDetails(int id);
+        Task<Bank> GetBankWithDetails(long id);
         Task<Bank> GetBankByName(string Name);
         Task<List<Bank>> GetAllBanksAsync();
+
+
     }
 }
